+++ R/prepInputs.R
@@ -395,7 +395,8 @@ prepInputs <- function(targetFile = NULL, url = NULL, archive = NULL, alsoExtrac
       messagePrepInputs("Loading object into R", verbose = verbose)
       if (identical(theFun, raster::raster) |
           identical(theFun, raster::stack) |
-          identical(theFun, raster::brick)) {
+          identical(theFun, raster::brick) |
+          identical(theFun, terra::rast)) {
         ## Don't cache the reading of a raster
         ## -- normal reading of raster on disk is fast b/c only reads metadata
         do.call(theFun, append(list(asPath(out$targetFilePath)), args))