+++ NEWS.md
@@ -4,11 +4,13 @@ Version 1.2.17
 ==============
 
 ## enhancements
-- `postProcess` now uses `terra` and `sf` by default throughout the family.
+- `postProcess` now uses `terra` and `sf` by default throughout the family. These are only activated by a user deciding to use the new family of functions (`postProcessTerra`, `cropTo`, `maskTo`, `projectTo`, `writeTo`) or by setting the `option(reproducible.useTerra = TRUE)`
+- new functions to assist with transition from `raster` to `terra` --> `maxFn` and `minFn`
 
 ## bugfixes
 - `Postgres` database backends were not working correctly because of a typo; fixed
 - When working with revdep `SpaDES.core`, there were some cases where the `Cache` was failing as it could not find the module name; fixed.
+- during transition from `postProcess` (using `raster` and `sp`) to `postProcessTerra`, some cases are falling through the cracks; these are being addressed.
 
 Version 1.2.16
 ==============